(PC-17814) identityCheck: separate api file into single files
