+++ src/features/identityCheck/api/logoutFromEduConnectIfAllowed.ts
@@ -1,0 +1,7 @@
+import { env } from 'libs/environment'
+
+export const logoutFromEduConnectIfAllowed = (logoutUrl: string | undefined) => {
+  if (logoutUrl && new RegExp(`^${env.EDUCONNECT_ALLOWED_DOMAIN}`, 'i').test(logoutUrl)) {
+    globalThis.window.open(logoutUrl)
+  }
+}
