+++ src/features/identityCheck/api/types.ts
@@ -1,0 +1,4 @@
+export interface MutationOptions {
+  onSuccess: () => void
+  onError: (error: unknown) => void
+}
