+++ src/features/identityCheck/constants.ts
@@ -1,0 +1,3 @@
+import { WEBAPP_V2_URL } from 'libs/environment'
+
+export const REDIRECT_URL_UBBLE = `${WEBAPP_V2_URL}/verification-identite/fin`
