+++ src/features/identityCheck/pages/identification/ubble/UbbleWebview.tsx
@@ -4,7 +4,8 @@ import { WebView } from 'react-native-webview'
 import styled from 'styled-components/native'
 
 import { IdentityCheckMethod } from 'api/gen'
-import { REDIRECT_URL_UBBLE, useIdentificationUrl } from 'features/identityCheck/api/api'
+import { useIdentificationUrl } from 'features/identityCheck/api/useIdentificationUrl'
+import { REDIRECT_URL_UBBLE } from 'features/identityCheck/constants'
 import { parseUrlParams } from 'features/identityCheck/utils/parseUrlParams'
 import { navigateToHome } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
