+++ src/features/identityCheck/pages/identification/ubble/UbbleWebview.web.tsx
@@ -2,7 +2,8 @@ import { useNavigation } from '@react-navigation/native'
 import React, { useEffect } from 'react'
 
 import { IdentityCheckMethod } from 'api/gen'
-import { REDIRECT_URL_UBBLE, useIdentificationUrl } from 'features/identityCheck/api/api'
+import { useIdentificationUrl } from 'features/identityCheck/api/useIdentificationUrl'
+import { REDIRECT_URL_UBBLE } from 'features/identityCheck/constants'
 import { navigateToHome } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { analytics } from 'libs/firebase/analytics'
