+++ src/features/identityCheck/pages/phoneValidation/CodeNotReceivedModal.tsx
@@ -3,10 +3,8 @@ import React, { FunctionComponent, useCallback } from 'react'
 import styled from 'styled-components/native'
 
 import { ApiError, extractApiErrorMessage } from 'api/apiHelpers'
-import {
-  usePhoneValidationRemainingAttempts,
-  useSendPhoneValidationMutation,
-} from 'features/identityCheck/api/api'
+import { usePhoneValidationRemainingAttempts } from 'features/identityCheck/api/usePhoneValidationRemainingAttempts'
+import { useSendPhoneValidationMutation } from 'features/identityCheck/api/useSendPhoneValidationMutation'
 import { useSubscriptionContext } from 'features/identityCheck/context/SubscriptionContextProvider'
 import { formatPhoneNumberWithPrefix } from 'features/identityCheck/pages/phoneValidation/utils'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
