+++ src/features/identityCheck/pages/phoneValidation/SetPhoneValidationCode.native.test.tsx
@@ -22,10 +22,8 @@ jest.mock('features/identityCheck/context/SubscriptionContextProvider', () => ({
   }),
 }))
 
-jest.mock('features/identityCheck/api/api', () => {
-  const ActualIdentityCheckAPI = jest.requireActual('features/identityCheck/api/api')
+jest.mock('features/identityCheck/api/usePhoneValidationRemainingAttempts', () => {
   return {
-    ...ActualIdentityCheckAPI,
     usePhoneValidationRemainingAttempts: jest.fn().mockReturnValue({
       remainingAttempts: 5,
       counterResetDatetime: 'time',
