+++ src/features/identityCheck/pages/phoneValidation/errors/PhoneValidationTooManySMSSent.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components/native'
 
-import { usePhoneValidationRemainingAttempts } from 'features/identityCheck/api/api'
+import { usePhoneValidationRemainingAttempts } from 'features/identityCheck/api/usePhoneValidationRemainingAttempts'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { timeDiffInHours } from 'libs/dates'
 import { plural } from 'libs/plural'
